extract `LincheckClassFileTransformer` into separate file
